requirements.txt
# EyeCare AI Agent - Requirements
# Core UI Framework
customtkinter>=5.2.0
Pillow>=10.0.0

# Computer Vision & Image Processing
opencv-python>=4.8.0
numpy>=1.24.0
mediapipe>=0.10.0

# AI & HTTP Clients
httpx>=0.25.0
openai>=1.0.0

# System & Hardware
pyserial>=3.5
screen-brightness-control>=0.20.0

# Utilities
pytz>=2023.3
plyer>=2.1.0
pystray>=0.19.0
screeninfo>=0.8.1
schedule>=1.2.0
python-dotenv>=1.0.0

# Data & Storage
sqlite-utils>=3.35.0

# Audio
pygame>=2.5.0

# Optional Hardware Sensors (install if you have sensors)
# adafruit-circuitpython-tsl2591>=1.3.0
# adafruit-blinka>=8.0.0

# Development Tools
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
